+++ android/src/main/java/com/twiliovoicereactnative/Constants.java
@@ -4,6 +4,7 @@ public class Constants {
   public static final String CALL_SID_KEY = "CALL_SID";
   public static final String VOICE_CHANNEL_LOW_IMPORTANCE = "notification-channel-low-importance";
   public static final String VOICE_CHANNEL_HIGH_IMPORTANCE = "notification-channel-high-importance";
+  public static final String VOICE_CHANNEL_DEFAULT_IMPORTANCE = "notification-channel-normal-importance";
   public static final String INCOMING_CALL_INVITE = "INCOMING_CALL_INVITE";
   public static final String CANCELLED_CALL_INVITE = "CANCELLED_CALL_INVITE";
   public static final String NOTIFICATION_ID = "NOTIFICATION_ID";
@@ -17,7 +18,7 @@ public class Constants {
   public static final String ACTION_FCM_TOKEN = "ACTION_FCM_TOKEN";
   public static final String ACTION_OUTGOING_CALL = "ACTION_OUTGOING_CALL";
   public static final String ACTION_PUSH_APP_TO_FOREGROUND = "ACTION_PUSH_APP_TO_FOREGROUND";
-
+  public static final String ACTION_PUSH_APP_TO_FOREGROUND_AND_MINIMIZE_NOTIFICATION = "ACTION_PUSH_APP_TO_FOREGROUND_AND_MINIMIZE_NOTIFICATION";
   public static final String FCM_TOKEN = "FCM_TOKEN";
   public static final String UUID = "UUID";
   public static final String NOTIFICATION_BANNER = " audio call";